CacheService getest, nog een test te gaan in ieder geval

+++ api/src/main/scala/com/leagueprojecto/api/Startup.scala
@@ -40,8 +40,8 @@ object Startup extends App with JsonProtocols {
   val matchHistoryService: ActorRef = system.actorOf(MatchHistoryService.props)
 
   // Service caches
-  val summonerCacheTime = config.getInt("riot.services.summonerbyname.cacheTime")
-  val matchhistoryCacheTime = config.getInt("riot.services.matchhistory.cacheTime")
+  val summonerCacheTime = config.getLong("riot.services.summonerbyname.cacheTime")
+  val matchhistoryCacheTime = config.getLong("riot.services.matchhistory.cacheTime")
   val cachedSummonerService: ActorRef     = system.actorOf(CacheService.props[Summoner](summonerService, summonerCacheTime))
   val cachedMatchHistoryService: ActorRef = system.actorOf(CacheService.props[List[MatchHistory]](matchHistoryService, matchhistoryCacheTime))
